Updates, bug fixes, add-ons
Learned about fetch.  In the initial response there are two properties (ok, statusCode) that provide understanding if the request was successful or not.  This will assist with keeping errors out of the console.

Improved the Espn models.  Switched their renderIds to nanoIds at 21 characters in length.  This will assist with duplicate Ids.  Created a class to properly maintain img tags.  All renderIds are logged under a window property now called renderCodes.  When a model renders on to the page the container will add an attribute called render-by containing the renderId.  Each model will also provide a look up method called (getByRenderCode).  This will allow you to easily find a model regardless of your location.

Added CustomEvents to ensure that each step of the round and game are performed and you can easily tap into the event.

Events that complete a full game are (onInit, onStart, onPlayerSelect, onRoundEnd, onGameOver, onLoadLocal, onSaveLocal)  They are triggered by changing the states of the GameController.